requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
